Reworking logic in DB, new views for mailers, quantity and active as row for product handling. Handling order to have products links, new routes restrictions

+++ app/controllers/products_controller.rb
@@ -3,23 +3,27 @@ class ProductsController < ApplicationController
   before_action :authorize_admin, only: [:create, :update, :destroy]
   before_action :set_product, only: [:show, :update, :destroy]
 
-  # GET /products
   def index
-    @products = Product.all
+    if current_user && current_user.admin?
+      products = Product.all
+    else
+      products = Product.all.where(active: true)
+    end
 
-    render json: @products.as_json(methods: :img_url)
+    if params[:name].present?
+      products = products.where("name ILIKE ?", "%#{params[:name]}%")
+    end
+
+    render json: products.as_json(methods: :img_url)
   end
 
-  # GET /products/1
   def show
     @product = Product.find(params[:id])
     render json: @product.as_json(methods: :img_url)
   end
 
-  # POST /products
   def create
     @product = Product.new(product_params)
-
     if @product.save
       render json: @product, status: :created, location: @product
     else
@@ -27,7 +31,6 @@ class ProductsController < ApplicationController
     end
   end
 
-  # PATCH/PUT /products/1
   def update
     if @product.update(product_params)
       render json: @product
@@ -36,30 +39,34 @@ class ProductsController < ApplicationController
     end
   end
 
-  # DELETE /products/1
-# DELETE /products/1
   def destroy
-    if @product.destroy
-      head :no_content
+    if @product.destroyable?
+      if @product.destroy
+        head :no_content
+      else
+        render json: { error: 'Une erreur a eu lieu lors de la suppression' }, status: :unprocessable_entity
+      end
     else
-      render json: { error: 'Une erreur a eu lieu lors de la suppression' }, status: :unprocessable_entity
+      @product.active = !@product.active
+      if @product.save
+        render json: { message: 'Le statut du produit a été mis à jour' }, status: :ok
+      else
+        render json: { error: 'Impossible de mettre à jour le statut du produit' }, status: :unprocessable_entity
+      end
     end
   end
 
-
   private
-    # Use callbacks to share common setup or constraints between actions.
     def set_product
       @product = Product.find(params[:id])
     end
 
-    # Only allow a list of trusted parameters through.
     def product_params
-      params.require(:product).permit(:name, :image, :price, :description, :origin, :variety, :categorie)
+      params.require(:product).permit(:name, :image, :price, :description, :origin, :variety, :category)
     end
     
     def authorize_admin
-      unless params[:isAdmin] == "true"
+      unless current_user && current_user.admin?
         render json: { error: 'Vous n\'avez pas accès à cette ressource' }, status: :unauthorized
       end
     end
